requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
falcon==3.0.0
greenlet==1.0.0
gunicorn==20.1.0
jwcrypto==0.8
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.4
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.8
sqlparse==0.4.1
structlog==21.1.0
tabulate==0.8.9
yoyo-migrations==7.3.1
argon2-cffi==20.1.0
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
falcon==3.0.0
greenlet==1.0.0
gunicorn==20.1.0
jwcrypto==0.8
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.4
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.8
sqlalchemy-mixins==1.3
sqlparse==0.4.1
structlog==21.1.0
tabulate==0.8.9
yoyo-migrations==7.3.1

boto3~=1.17.84
botocore~=1.20.84